LibTLS: Move `DefaultRootCACertificates` to own file
